refactor: move type definitions
